feat: characters CRUD
This commit migrates the Characters CRUD from OpenRPG to OpenRPG-UI. The
least amount of changes to the .vue components want to be made, just to
evaluate how far we can go copy and pasting .vue files from OpenRPG
without having to rewrite work.

Copies the following components:

* CharacterTable.vue
* alerts/SuccessMessage.vue
* alerts/ValidationErrors.vue

Additionally it copies the following views:

* characters/index.vue => pages/characters/index.vue
* characters/edit.vue => pages/characters/_id/edit.vue
* characters/show.vue => pages/characters/_id/index.vue
* characters/create.vue => pages/characters/create.vue

Plus the schema, moved from
views/characters/schemas/character-schema.vue to schemas/character.js.

VueFormGenerator is installed as a dependency and it works good.

The following changes had to be made:

* nuxt-eslint refuses to serve some components until they get their
  errors fixed.
* axios has to be rewrited to this.$axios, because it is a plugin inside
  the Nuxt app (which has the this context).
* Routes have their name changed due to the way Nuxt works.