feat: make the API callable from Nuxt
This commit configures a Nuxt proxy at /api, pointing to
http://localhost:8000/api, so that calls can be made from Axios to
/api/characters. Nuxt proxy will send them to
http://localhost:8000/api/characters. This avoids CORS issues.

Because requests have to be authenticated, I've added a plugin that sets
at the application boot a Bearer token so that requests don't fail. I've
opened https://github.com/rubendelblanco/openrpg/issues/49 to evaluate a
better way to provide a login flow to the final player.