feat: port Users CRUD
This commit will migrate the Users CRUD in OpenRPG to openrpg-ui. The
UserTable component has been migrated. It had to be changed the syntax
for the slot in use to present the action buttons in the table, but
aside from that, no further changes had to be made for it to work.