fix: backport JSON:API changes made to backend
This commit will change the frontend API client so that it fetches data
about characters and users using the new keys present in the API sent by
the backend.

The code has a lot of things to improve, but I've refrained from adding
these changes to the commit, which only reflects the minimum viable
patch to make the frontend work again.