fix(test): pull valid ava config
It seems that the configuration file for ava has changed format at some
point, and the project generated with create-nuxt-app a few months ago
does not include a valid ava config anymore.

This commit fixes that by pulling the bootstrapped ava config files and
test/ directory that are generated when an updated release of
create-nuxt-app is run generating a project that uses AVA as the test
runner.