fix(test): provide valid test cases
This commit removes the Logo.spec.js spec, which is not valid anymore
due to the way things have changed in this repository, with a valid test
case that validates the SuccessMessage.

A few snapshot test cases are generated, because I bet on snapshot tests
to quickly catch regressions introduced into the user interface
accidentally.