feat(dev): install @storybook/vue
This commit installs Storybook. Storybook is an UI component explorer
that allows developers to quickly build a canvas where Vue components
can be imported and rendered outside of the application, to make them
easier to validate by designers and testers and other developers, if the
components are reusable enough.

This commit also adds a Storybook story to validate the look and feel of
the SuccessMessage, as a demo to showcase what Storybook is able to do,
and to validate the layout of the component.