feat(dev): install @storybook/addon-actions
addon-actions is a Storybook addon that allows the Storybook UI to
present dispatched actions under certain conditions. For instance,
addon-actions can be used to view in the web browser an event emitted by
a component, if the event handler dispatches a Storybook action.

This commit also adds a Storybook story that emits an action when a
SuccessMessage is clicked, as a demo to showcase how to handle Vue
events dispatching Storybook actions.