feat(dev): install @storybook/addon-knobs
addon-knobs is a Storybook addon that puts an additional layer of
settings into the Storybook web application, so that the user can live
update on their web browser dynamic data that will be used to feed
component props.

This commit also adds a Storybook story that allows the user to put
custom props into a SuccessMessage, as a demo to showcase how to provide
custom props into a component of this application.