fix(deps): downgrade to core-js@2 (#74)
This pull request will fix a regression introduced eventually into the
project due to some package update, that causes the application to not
start, because Nuxt is not compatible with core-js@3 at the moment.

Additionally, this pull request introduces a GitHub action that attempts
to run `yarn build` in order to make sure that the project is at least
compilable. This is a smoke test that should not replace endpoint
testing to assert that at least some pages work.