fix: make VueRadialMenu support SSR
Due to the way vue-radial-menu is packed, you cannot use at the moment
SSR when rendering this component because the import of the minified
bundle fails.

Fortunately, the radial menu is locally refactored into a wrapping
component, so this can be fixed by importing the RadialMenu component
from the src/ directory of the library instead.