feat: change the build mode to universal (#75)
This commit will set by default the application mode to universal,
previously set to spa. The reason it was set to spa is because
vue-radial-menu is not compatible with SSR when using the minified dist/
bundle, because it somehow packs too much dependencies.

However, if instead of importing vue-radial-menu (which imports the
whole bundle at dist/radialMenu.common.js), each .vue file is imported
separately, the application works fine using server side rendering.