feat(ops): add Dockerfile
This commit adds a Dockerfile that allows OpenRPG-UI to be deployed into
a Docker container. This is the most simple Dockerfile that I could
achieve, which packs the application and runs it on server side mode,
now that it is possible to do so.

It was a requirement to be able to run the server in SSR mode, because
it is the only way that allows Nuxt proxy to be set, making easier to
execute API calls, because they are requested by the web browser to
Nuxt, avoiding cross-origin issues that usually happen, without the need
to install CORS headers.

Because Docker containers usually run isolated, additional environment
variables were added to nuxt.config.js to fine tune the way Nuxt proxy
finds the OpenRPG backend server. These changes have been documented in
the README.md file.

Additionally, to quickly deploy these images, additional GitHub Actions
have been imported into this project, pushing into GitHub Packages the
Docker image in master on each merge of a branch. These actions are
taken from https://github.com/makigas/clank/tree/master/.github/workflows,
where they have been working quite well for a long time.