fix(test): fix test suites (#92)
Tests are fixed.

Uses Cypress rather than Ava for the E2E tests to have more control over
the spawning of the server and because Ava doesn't seem to be fully
designed to run E2E tests, it is just duck tape around the programmatic
Nuxt.js API and using jsdom as a parser.

With E2E outside of Ava, I refactor the structure of the unit test suite
so that it is the single kind of tests run.

Also, a few tests are added to actually prove that E2E and unit tests
are working, including snapshot testing, and GitHub Actions are
introduced to make sure that regressions with the test suite are caught
in the future. The idea behind snapshot testing is that, if a
bootstrap-vue upgrade breaks the component, it will be noticeable due
to the snapshots mismatching.